feat(http): support URL encoding for forms
Add the `form_encoding` parameter to HTTP methods. When it is set to
`application/x-www-form-urlencoded`, encode form data as a string of
key-value pairs. This is the same encoding as Go's `http.PostForm`.

URL encoding is now the default for `http.post()` and other methods
that accept a form body.