Series data structure represents a DataFrame column
Implement a Series. For now, only three types of data are allowed: ints, floats,and strings (objects). Indexes, names, and data types are also here. The only things a Series can do for now are being constructed, being printed, and getting a cell from either the numerical index, or the name from the index.