feat(dataframe): Index and basic DataFrame with columns
Index is used to store axis labels for Series and DataFrame. It has a name, and can be stringified. The text values in an index can be manipulated using stringMethods. A Series with an Index with a name will stringify with that name.

DataFrame exists, but only has a bit of functionality. It has columns, which are an index. It has an index which stringifies. DataFrame columns can be get'ed or set'ed, using Series objects. DataFrame can be constructed using a list or dict.

Build the series contents using a builder that handles all type coercion. This allows DataFrame to retain its actual input types. Series now handles boolean types.