feat(dataframe): parse_csv to parse csv from text
Prior to this, read_csv would parse csv from text. This was different from the pandas.DataFrame implementation, which would require a url to be passed to read_csv. This difference caused confusion. In the future, once the story around http access is fixed, the read_csv will be restored to match the pandas version.