fix(dataframe): Invariant that body is tabular, OutputConfig exists
In order to avoid potential segfaults, we must ensure that DataFrames always have a tabular body; that each column has the same height. Use a single constructor everywhere that enforces this rule, as well as the existence of an OutputConfig. Update the append method so that rows of different sizes are handled correctly, by padding them with None values.