refactor(dataframe): Separate Index type from implementations
Pandas has multiple types of Indexes. An Index can be a list of strings, or a range of numbers, or a list of numbers. We want to use strong typing for the DataFrame object, so turn Index into a single type (getting rid of the old RangeIndex), and use an interface to represent these different types, with an implementation for each. This will bring starlark dataframe's behavior to be closer to the python pandas version.